Move files to proper locations 2
