+++ src/TSTransformer/nodes/transformMethodDeclaration.ts
